+++ components/NodeContent/src/NodeContent.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { styled } from '@hautechai/webui.themeprovider';
 import { themeVars } from '@hautechai/webui.themeprovider';
+import { Divider } from '@hautechai/webui.divider';
 
 export interface NodeContentProps {
     children: React.ReactNode;
@@ -15,5 +16,10 @@ const Container = styled.div`
 `;
 
 export const NodeContent: React.FC<NodeContentProps> = ({ children, testId }) => {
-    return <Container data-testid={testId}>{children}</Container>;
+    return (
+        <>
+            <Container data-testid={testId}>{children}</Container>
+            <Divider />
+        </>
+    );
 };