+++ components/Popover/src/Popover.tsx
@@ -2,6 +2,12 @@ import { forwardRef, useCallback, useImperativeHandle, useState } from 'react';
 import { Popover as TinyPopover } from 'react-tiny-popover';
 import { PopoverProps, PopoverRef } from './Popover.types';
 import S from './Popover.styles';
+import { styled } from '@hautechai/webui.themeprovider';
+
+const ChildrenContainer = styled.div`
+    display: flex;
+    flex: 1;
+`;
 
 export const Popover = forwardRef<PopoverRef, PopoverProps>((props: PopoverProps, ref) => {
     const [isOpen, setOpen] = useState(false);
@@ -17,15 +23,15 @@ export const Popover = forwardRef<PopoverRef, PopoverProps>((props: PopoverProps
     }));
 
     return (
-        <div data-testid={props.testId}>
-            <TinyPopover
-                content={<S.Container>{props.content({ close })}</S.Container>}
-                isOpen={isOpen}
-                onClickOutside={close}
-                positions={props.contentPositions ?? ['top', 'bottom', 'left', 'right']}
-            >
-                <div onClick={toggle}>{props.trigger()}</div>
-            </TinyPopover>
-        </div>
+        <TinyPopover
+            content={<S.Container>{props.content({ close })}</S.Container>}
+            isOpen={isOpen}
+            onClickOutside={close}
+            positions={props.contentPositions ?? ['top', 'bottom', 'left', 'right']}
+        >
+            <ChildrenContainer onClick={toggle} data-testid={props.testId}>
+                {props.trigger()}
+            </ChildrenContainer>
+        </TinyPopover>
     );
 });