+++ components/Menu/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.13](https://github.com/HautechAI/webui/compare/@hautechai/webui.menu@2.1.12...@hautechai/webui.menu@2.1.13) (2025-09-14)
+
+**Note:** Version bump only for package @hautechai/webui.menu
+
 ## [2.1.12](https://github.com/HautechAI/webui/compare/@hautechai/webui.menu@2.1.11...@hautechai/webui.menu@2.1.12) (2025-09-06)
 
 **Note:** Version bump only for package @hautechai/webui.menu