+++ components/NodeContent/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.10](https://github.com/HautechAI/webui/compare/@hautechai/webui.nodecontent@1.0.9...@hautechai/webui.nodecontent@1.0.10) (2025-09-15)
+
+**Note:** Version bump only for package @hautechai/webui.nodecontent
+
 ## [1.0.9](https://github.com/HautechAI/webui/compare/@hautechai/webui.nodecontent@1.0.8...@hautechai/webui.nodecontent@1.0.9) (2025-09-06)
 
 **Note:** Version bump only for package @hautechai/webui.nodecontent